requirements.txt
Flask
Gunicorn 
python-dotenv
pymongo
requests
vosk
werkzeug